+++ src/VbrMigrationPrecheck/Public/Invoke-VbrMigrationPrecheck.ps1
@@ -78,7 +78,8 @@ function Invoke-VbrMigrationPrecheck {
         'Test-AgentVersions'
         'Test-AgentDisabledPolicies'
         'Test-MacAgentDomainAuth'
-        'Test-ProtectionGroupPostMigration'
+        'Test-ProtectionGroupPostMigration',
+        'Test-AdProtectionGroupLdapPort'
         'Test-NetAppOntapRole'
         'Test-StoragePluginVersions'
         'Test-NimbleFips'